chore: folder restructuring by features
