+++ ichub-frontend/src/features/catalog-management/components/product-detail/InstanceProductsTable.tsx
