+++ ichub-frontend/src/features/catalog-management/components/product-list/CardChip.tsx
