+++ ichub-frontend/src/features/catalog-management/components/product-list/ProductCard.tsx
